==> Script_4_Further_Fidel.Rmd <==
---
title: "Further fidelity"
output: 
  rmdformats::material:
    highlight: kate
    css: web_style.css
    thumbnails: false
    lightbox: true
    gallery: true
    cards: true
    self_contained: no
    number_sections: no
    code_folding: hide
    fig_caption: yes
---

```{r setup, include = FALSE, message = FALSE, warning = FALSE, echo=FALSE}

# Set general settings for Markdown file 
  options(max.print="75")
  knitr::opts_chunk$set(echo=TRUE,
  	             #cache=TRUE,
                 prompt=FALSE,
                 tidy=TRUE,
                 comment=NA,
                 message=FALSE,
                 warning=FALSE,
                 results = FALSE)
  knitr::opts_knit$set(width=75)
  
# Load packages
  library(afex)
  library(expss)
  library(ggwordcloud)
  library(gtsummary)
  library(jmv)
  library(kableExtra)
  library(miceadds)
  library(pander)
  library(plyr)
  library(rmdformats)
  library(sjPlot)
  library(tadaatoolbox)
  library(tidyverse)
  library(tm)
  library(wordcloud)
  library(XLConnect)
  
# Swipe environment
  rm(list=ls())
  
  
# Set figure theme  
  theme_SN = theme(axis.title.y = element_text(size = 15,
                                               margin = margin(t = 0, r = 20, b = 0, l = 0)),
          panel.grid.major.x = element_blank(), panel.grid.minor.x = element_blank(),
          panel.grid.major.y = element_line(colour = "black", linetype = "dotted", size=0.6),
          panel.grid.minor.y = element_blank(),
          panel.border = element_blank(),
          text=element_text(size = 15),
          legend.position = "none")

```

```{r load_data, include = FALSE}
# Load app questionnaire data
  load.Rdata(filename="./data/app_qn.Rdata", "app_qn")
  load.Rdata(filename="./data/qn_data.Rdata", "qn_data")
```

# Further fidelity measures {.tabset .tabset-pills}

## Motivation

```{r fidel_stats_prep, include = FALSE}

# Select data for acceptibility questions
  app_stats = as_tibble(app_qn)
  app_stats = app_stats %>% dplyr::select(Fun, Motivation, Practiced_without_parent, 
                                          Use_App_self_explanatory, Use_Training_daily_life,
                                          Behav_Dealing_w_Feelings, Behav_Interest_Languages)
  
# Add training info
  app_stats$group = app_qn$App  
  
# Rename groups
  app_stats$group = dplyr::recode(app_stats$group, 
                         CT = "Controls", ZE = "Zirkus Empathico")
  
  app_stats[app_stats$group == "0",] = NA
  
  app_stats = app_stats[complete.cases(app_stats$group),]
  

```

```{r fidel_plot, fig.width = 8, fig.height = 4}

# https://cran.r-project.org/web/packages/sjPlot/vignettes/plot_likert_scales.html

# Recode variable 
  app_stats_lab  = data.frame(lapply(app_stats, factor, ordered=TRUE, 
                          levels=1:5, 
                          labels=c("Do not agree", "Rather not agree", "Partly agree","Rather agree","Agree")))

  app_stats_lab$group = app_stats$group

# Separate by training / control group
  app_stats_ZE = app_stats_lab[app_stats_lab$group == "Zirkus Empathico",]
  app_stats_SB = app_stats_lab[app_stats_lab$group == "Controls",]
  
  app_stats_ZE = subset(app_stats_ZE, select = -c(group))
  app_stats_SB = subset(app_stats_SB, select = -c(group))


# Create main plot  
   
  ZE_fidel = plot_likert(app_stats_ZE, catcount = 5, values = FALSE, wrap.legend.labels = 5,
                geom.colors = c("#f0c99d","#e1943a","#d97909","#6d3d05","#2b1802"),
                rel_heights = c(8, 12),
                title = "Zirkus Empathico",
                wrap.labels = 65,
                reverse.scale = TRUE,
                show.n = FALSE,
                axis.labels = c("","","","","","","",""))+
                theme_bw() +
                theme(legend.position = "bottom", panel.grid.major.x = element_line(color = "grey"),
                   panel.grid.major.y = element_line(colour = "grey"),
                   panel.border = element_blank(),
                   axis.ticks.y = element_blank(), axis.ticks.x = element_blank(),
                   axis.text.x = element_text(size = 5),
                   legend.text = element_text(size = 4),
                   legend.key.size = unit(0.3, "cm"))
    
    
  CT_fidel = plot_likert(app_stats_SB, catcount = 5, values = FALSE, wrap.legend.labels = 5,
                geom.colors = c("#dbdbdb","#b7b7b7","#878787","#515151","#1b1b1b"),
                rel_heights = c(8, 12),
                title = "Controls",
                wrap.labels = 65,
                reverse.scale = TRUE,
                show.n = FALSE,
                legend.labels = NULL,
                axis.labels = c("My child enjoyed the training.",
                                "My child was motivated to do the training.",
                                "My child trained mostly without me.",
                                "The training was self-explanatory.",
                                "The training was compatible with daily routines.",
                                "My child copes better with its own feelings.",
                                "My child is more interested in other languages.")) +
            theme_bw() +
            theme(legend.position = "bottom", panel.grid.major.x = element_line(color = "grey"),
                   panel.grid.major.y = element_line(colour = "grey"),
                   panel.border = element_blank(),
                   axis.ticks.y = element_blank(), axis.ticks.x = element_blank(),
                   axis.text.x = element_text(size = 5),
                   legend.text = element_text(size = 4),
                   legend.key.size = unit(0.3, "cm"))

  fig_fidel = cowplot::plot_grid(CT_fidel, ZE_fidel, nrow = 1, rel_widths=c(1, .45))
  
  fig_fidel

  
```

```{r fidel_stats, results = "asis"}


# Apply labels
  app_stats = apply_labels(app_stats,
                      Fun = "My child enjoyed the training.",
                      Motivation = "My child was motivated to do the training.",
                      Practiced_without_parent = "My child trained mostly without me.",
                      Use_App_self_explanatory = "The training was self-explanatory.",
                      Use_Training_daily_life = "The training was compatible with daily routines.",
                      Behav_Dealing_w_Feelings = "My child copes better with its own feelings.",
                      Behav_Interest_Languages = "My child is more interested in other languages")

# Prepare table
  app_stats_table =
    tbl_summary(
      app_stats ,
      by = group, # split table by group
      type = c(Fun, Motivation, Practiced_without_parent, Use_App_self_explanatory,
                Use_Training_daily_life, Behav_Dealing_w_Feelings, Behav_Interest_Languages) ~ "continuous",
      statistic = list(all_continuous() ~ "{mean} ({sd})", # descriptives definition
                     all_categorical() ~ "{n} / {N} ({p}%)"),
      digits = all_continuous() ~ 2,
      missing = "no" # don't list missing data separately
    ) %>%
    add_n() %>% # add column with total number of non-missing observations
    add_p() %>% # test for a difference between groups
    modify_header(label = "**Variable**") %>% # update the column header
    bold_labels()  
  
  
# Print table
  app_stats_table
  
```

*Please note: 2 parental ratings are missing for this fidelity rating.

## Acceptance

**My child had / was...**

```{r app_acc, results = "asis"}

# Select data for acceptibility questions
  app_acc = as_tibble(app_qn)
  app_acc = app_acc %>% dplyr::select(Fun:Practiced_without_parent)
  
# Recode variable 
  app_acc  = data.frame(lapply(app_acc , factor, ordered=TRUE, 
                          levels=1:5, 
                          labels=c("Do not agree", "Rather not agree", "Partly agree","Rather agree","Agree")))
  
# Add training info
  app_acc$group = app_qn$App  

# Separate by training / control group
  app_acc_ZE = app_acc[app_acc$group == "ZE",]
  app_acc_SB = app_acc[app_acc$group == "CT",]
  
  app_acc_ZE = subset(app_acc_ZE, select = -c(group))
  app_acc_SB = subset(app_acc_SB, select = -c(group))

# Combine data sets (account for different lengths of data frames)
  comb_acc_app = merge(app_acc_SB, app_acc_ZE, by = "row.names", all = T, suffixes = c("",""))
  comb_acc_app$Row.names <- NULL

# Plot likert plot
  plot_likert(comb_acc_app, catcount = 5, values = FALSE, wrap.legend.labels = 5, 
                c(rep("Controls", 8), rep("Zirkus Empathico", 8)),
                geom.colors = c("#040a0b","#1a3e43","#2b6770","#80a4a9","#bfd1d4"),
                rel_heights = c(9, 11),
                wrap.labels = 65,
                reverse.scale = TRUE,
                show.n = FALSE,
                axis.labels = c("fun engaging with the app.", "motivated to train with the app.",
                                "frustrated during training.", "bored during training.",
                                "wanted to play longer than agreed.",
                                "motivated by the app's rewards.",
                                "difficulty understanding some of the app's content.",
                                "trained mostly without me."))

```

```{r wordcloud_prep}

# http://www.sthda.com/english/wiki/text-mining-and-word-cloud-fundamentals-in-r-5-simple-steps-you-should-know
# https://cran.r-project.org/web/packages/ggwordcloud/vignettes/ggwordcloud.html

# Read txt files
  text_CG_likes = readLines("./data/wc_CG_likes.txt")
  text_CG_dislikes = readLines("./data/wc_CG_dislikes.txt")
  text_TG_likes = readLines("./data/wc_TG_likes.txt")
  text_TG_dislikes = readLines("./data/wc_TG_dislikes.txt")
  
# Load data as corpus
  CG_likes = Corpus(VectorSource(text_CG_likes))
  CG_dislikes = Corpus(VectorSource(text_CG_dislikes))
  TG_likes = Corpus(VectorSource(text_TG_likes))
  TG_dislikes = Corpus(VectorSource(text_TG_dislikes))

```

**What did your child like most about digital training?**

```{r wordcloud_game_likes}

# Build a term-document matrix
  dtm = TermDocumentMatrix(CG_likes)
  m = as.matrix(dtm)
  v = sort(rowSums(m),decreasing=TRUE)
  d_CG_likes = data.frame(word = names(v),freq=v)

  dtm = TermDocumentMatrix(TG_likes)
  m = as.matrix(dtm)
  v = sort(rowSums(m),decreasing=TRUE)
  d_TG_likes = data.frame(word = names(v),freq=v)
  
# Build word coulds
  CG_likes_word = ggplot(d_CG_likes, aes(label = word, size = freq, color = freq))+
    geom_text_wordcloud(area_corr = TRUE, rm_outside = TRUE) +
    scale_size_area(max_size = 22) +
    theme_minimal() +
    scale_color_gradient(low = "#041012", high = "#5fa7b3")+
    ggtitle ("Controls")

  TG_likes_word = ggplot(d_TG_likes, aes(label = word, size = freq, color = freq)) +
    geom_text_wordcloud(area_corr = TRUE,  rm_outside = TRUE) +
    scale_size_area(max_size = 22) +
    theme_minimal() +
    scale_color_gradient(low = "#63430a", high = "#c48310")+
    ggtitle ("Zirkus Empathico")

# Display plots
  fig_like_words = cowplot::plot_grid(CG_likes_word, TG_likes_word, nrow = 2, rel_widths=c(1, 1))
  fig_like_words 

```

<br>

**What did he/she like less?**

```{r wordcloud_game_dislikes}

# Build a term-document matrix
  dtm = TermDocumentMatrix(CG_dislikes)
  m = as.matrix(dtm)
  v = sort(rowSums(m),decreasing=TRUE)
  d_CG_dislikes = data.frame(word = names(v),freq=v)

  dtm = TermDocumentMatrix(TG_dislikes)
  m = as.matrix(dtm)
  v = sort(rowSums(m),decreasing=TRUE)
  d_TG_dislikes = data.frame(word = names(v),freq=v)
  
# Build word coulds
  CG_dislikes_word = ggplot(d_CG_dislikes, aes(label = word, size = freq, color = freq))+
    geom_text_wordcloud(area_corr = TRUE, rm_outside = TRUE) +
    scale_size_area(max_size = 21) +
    theme_minimal() +
    scale_color_gradient(low = "#041012", high = "#5fa7b3")+
    ggtitle ("Controls")

  TG_dislikes_word = ggplot(d_TG_dislikes, aes(label = word, size = freq, color = freq)) +
    geom_text_wordcloud(area_corr = TRUE, rm_outside = TRUE) +
    scale_size_area(max_size = 21) +
    theme_minimal() +
    scale_color_gradient(low = "#63430a", high = "#c48310")+
    ggtitle ("Zirkus Empathico")

# Display plots
  fig_like_words = cowplot::plot_grid(CG_dislikes_word, TG_dislikes_word, nrow = 2, rel_widths=c(1, 1))
  fig_like_words 
  
  
  
  

```


**What did your child find difficult while exercising? What did he or she possibly not understand?**

```{r wordcloud_diff}

# Read txt files
  text_CG_diff = readLines("./data/wc_CG_diff.txt")
  text_TG_diff = readLines("./data/wc_TG_diff.txt")
  
# Load data as corpus
  CG_diff = Corpus(VectorSource(text_CG_diff))
  TG_diff = Corpus(VectorSource(text_TG_diff))

# Build a term-document matrix
  dtm = TermDocumentMatrix(CG_diff)
  m = as.matrix(dtm)
  v = sort(rowSums(m),decreasing=TRUE)
  d_CG_diff = data.frame(word = names(v),freq=v)

  dtm = TermDocumentMatrix(TG_diff)
  m = as.matrix(dtm)
  v = sort(rowSums(m),decreasing=TRUE)
  d_TG_diff = data.frame(word = names(v),freq=v)
  
# Build word coulds
  CG_diff_word = ggplot(d_CG_diff, aes(label = word, size = freq, color = freq))+
    geom_text_wordcloud(area_corr = TRUE) +
    scale_size_area(max_size = 18) +
    theme_minimal() +
    scale_color_gradient(low = "#041012", high = "#5fa7b3")+
    ggtitle ("Controls")

  TG_diff_word = ggplot(d_TG_diff, aes(label = word, size = freq, color = freq)) +
    geom_text_wordcloud(area_corr = TRUE) +
    scale_size_area(max_size = 18) +
    theme_minimal() +
    scale_color_gradient(low = "#63430a", high = "#c48310")+
    ggtitle ("Zirkus Empathico")

# Display plots
  fig_diff_words = cowplot::plot_grid(CG_diff_word, TG_diff_word, nrow = 2, rel_widths=c(1, 1))
  fig_diff_words 

```

## Usability 

```{r app_usa}

# Select data for acceptibility questions
  app_usa = as_tibble(app_qn)
  app_usa = app_usa %>% dplyr::select(Use_App_self_explanatory:Use_Training_Stressful)
  

# Recode variable 
  app_usa  = data.frame(lapply(app_usa , factor, ordered=TRUE, 
                          levels=1:5, 
                          labels=c("Do not agree", "Rather not agree", "Partly agree","Rather agree","Agree")))
  
# Add training info
  app_usa$group = app_qn$App  
  

# Separate by training / control group
  app_usa_ZE = app_usa[app_usa$group == "ZE",]
  app_usa_SB = app_usa[app_usa$group == "CT",]
  
  app_usa_ZE = subset(app_usa_ZE, select = -c(group))
  app_usa_SB = subset(app_usa_SB, select = -c(group))

# Combine data sets (account for different lengths of data frames)
  comb_int_use = merge(app_usa_SB, app_usa_ZE, by = "row.names", all = T, suffixes = c("",""))
  comb_int_use$Row.names <- NULL

# Plot likert plot
  plot_likert(comb_int_use, catcount = 5, values = FALSE, wrap.legend.labels = 5, 
                c(rep("Controls", 5), rep("Zirkus Empathico", 5)),
                geom.colors = c("#121117","#2e2b3b","#5b5675","#7c7891","#ceccd6"),
                rel_heights = c(8, 10),
                wrap.labels = 65,
                reverse.scale = TRUE,
                show.n = FALSE,
                axis.labels = c("App was self-explanatory.", "The training manual was helpful.",
                                "Training was compatible the daily routines.", "I felt adequately supervised.",
                                "I was mentally stressed during the training."))

```

## Perceived change after training

**After completion of the training, me as dad/mom found that...**

```{r app_behav}

# Select data for acceptibility questions
  app_beh = as_tibble(app_qn)
  app_beh = app_beh %>% dplyr::select(Behav_Interest_Feelings:Behav_Interest_Languages)

# Recode variable 
  app_beh  = data.frame(lapply(app_beh , factor, ordered=TRUE, 
                          levels=1:5, 
                          labels=c("Do not agree", "Rather not agree", "Partly agree","Rather agree","Agree")))

# Add training info
  app_beh$group = app_qn$App

# Separate by training / control group
  app_beh_ZE = app_beh[app_beh$group == "ZE",]
  app_beh_SB = app_beh[app_beh$group == "CT",]
  
  app_beh_ZE = subset(app_beh_ZE, select = -c(group))
  app_beh_SB = subset(app_beh_SB, select = -c(group))

# Combine data sets (account for different lengths of data frames)
  comb_int_use = merge(app_beh_SB, app_beh_ZE, by = "row.names", all = T, suffixes = c("",""))
  comb_int_use$Row.names <- NULL

# Plot likert plot
  plot_likert(comb_int_use, catcount = 5, values = FALSE, wrap.legend.labels = 5, 
                c(rep("Controls", 9), rep("Zirkus Empathico", 9)),
                geom.colors = c("#23180d","#583c22","#9e6c3c","#c09369","#d8bca1"),
                rel_heights = c(8, 10),
                wrap.labels = 65,
                reverse.scale = TRUE,
                show.n = FALSE,
                axis.labels = c("my child shows more interest in feelings.", "my child deals better with its own feelings.",
                                "my child reacts more appropriately to others's feelings.", "my child is more open-minded about his environment.",
                                "my child spends more time with other children.","communication about feelings in our family improved.",
                                "I have a better relationship with my child.","my child uses English words or phrases more often.","my child is more interested in other languages."))

```


**Has your child changed in terms of his or her behavior in school/kindergarten/family? If so, please describe how these changes manifested themselves.**

```{r wordcloud_change}

# Read txt files
  text_CG_chang = readLines("./data/wc_CG_chang.txt")
  text_TG_chang = readLines("./data/wc_TG_chang.txt")
  
# Load data as corpus
  CG_chang = Corpus(VectorSource(text_CG_chang))
  TG_chang = Corpus(VectorSource(text_TG_chang))

# Build a term-document matrix
  dtm = TermDocumentMatrix(CG_chang)
  m = as.matrix(dtm)
  v = sort(rowSums(m),decreasing=TRUE)
  d_CG_chang = data.frame(word = names(v),freq=v)
  
  dtm = TermDocumentMatrix(TG_chang)
  m = as.matrix(dtm)
  v = sort(rowSums(m),decreasing=TRUE)
  d_TG_chang = data.frame(word = names(v),freq=v)
  
  
# Build word coulds
  CG_chang_word = ggplot(d_CG_chang, aes(label = word, size = freq, color = freq))+
    geom_text_wordcloud(area_corr = TRUE, rm_outside = TRUE) +
    scale_size_area(max_size = 20) +
    theme_minimal() +
    scale_color_gradient(low = "#041012", high = "#5fa7b3")+
    ggtitle ("Controls")

  TG_chang_word = ggplot(d_TG_chang, aes(label = word, size = freq, color = freq)) +
    geom_text_wordcloud(area_corr = TRUE, rm_outside = TRUE) +
    scale_size_area(max_size = 20) +
    theme_minimal() +
    scale_color_gradient(low = "#63430a", high = "#c48310")+
    ggtitle ("Zirkus Empathico")

# Display plots
  fig_chang_words = cowplot::plot_grid(CG_chang_word, TG_chang_word, nrow = 2, rel_widths=c(1, 1))
  fig_chang_words   

```


## Parental involvement

<br>

**How was your involvement in your child's training (e.g. did your child tend to train alone or were you present the whole time, did you tend to observe or did you have to explain a lot)?**

```{r wordcloud_involv}

# Read txt files
  text_CG_involv = readLines("./data/wc_CG_involv.txt")
  text_TG_involv = readLines("./data/wc_TG_involv.txt")
  
# Load data as corpus
  CG_involv = Corpus(VectorSource(text_CG_involv))
  TG_involv = Corpus(VectorSource(text_TG_involv))

# Build a term-document matrix
  dtm = TermDocumentMatrix(CG_involv)
  m = as.matrix(dtm)
  v = sort(rowSums(m),decreasing=TRUE)
  d_CG_involv = data.frame(word = names(v),freq=v)
 
  dtm = TermDocumentMatrix(TG_involv)
  m = as.matrix(dtm)
  v = sort(rowSums(m),decreasing=TRUE)
  d_TG_involv = data.frame(word = names(v),freq=v)
  
# Build word coulds
  CG_involv_word = ggplot(d_CG_involv, aes(label = word, size = freq, color = freq))+
    geom_text_wordcloud(area_corr = TRUE, rm_outside = TRUE) +
    scale_size_area(max_size = 18) +
    theme_minimal() +
    scale_color_gradient(low = "#041012", high = "#5fa7b3")+
    ggtitle ("Controls")

  TG_involv_word = ggplot(d_TG_involv, aes(label = word, size = freq, color = freq)) +
    geom_text_wordcloud(area_corr = TRUE, rm_outside = TRUE) +
    scale_size_area(max_size = 18) +
    theme_minimal() +
    scale_color_gradient(low = "#63430a", high = "#c48310")+
    ggtitle ("Zirkus Empathico")

# Display plots
  fig_involv_words = cowplot::plot_grid(CG_involv_word, TG_involv_word, nrow = 2, rel_widths=c(1, 1))
  fig_involv_words 

```





# Session info

<!-- Provide session info  -->

```{r session_info, results = TRUE}

# Get session info 
  sessionInfo()

```
